requirements.txt
neptune==1.14.0
numpy==2.2.2
pandas==2.2.3
python-dotenv==1.1.0
scikit-learn==1.6.1
timm==1.0.15
torch==2.6.0+cu124
torchvision==0.21.0+cu124
tqdm==4.67.1
imagehash==4.3.2
